requirements.txt
# Core
fastapi==0.129.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.37

# Vector store
qdrant-client==1.16.2

# Embeddings and reranking
sentence-transformers==5.2.2

# BM25 keyword search
whoosh==2.7.4

# PDF parsing
pdfplumber==0.11.6

# LLM (new SDK — google-generativeai is deprecated)
google-genai==1.63.0

# Auth
pyjwt==2.10.1
